% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_predict.R
\name{append_abilities}
\alias{append_abilities}
\title{Appends the estimated abilities of each user in the training data to the training data}
\usage{
append_abilities(train_wide, mod)
}
\arguments{
\item{train_wide}{the wide training data (collapsed)}

\item{mod}{the fitted IRT model (to the training data)}
}
\value{
the wide training data with ability appended
}
\description{
given a fitted IRT model and the wide training data the model was fitted to
append the ability estimate for the user to the row
}
