% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_predict.R
\name{append_repeats}
\alias{append_repeats}
\title{Appends a repeat column to tell us whether repeats for an item have occurred}
\usage{
append_repeats(test_data, train_long)
}
\arguments{
\item{test_data}{the test data}

\item{train_long}{the UNCOLLAPSED training data}
}
\value{
the test data with a new column indicating if it's a repeat
}
\description{
given test and training data, looks to see if the observation in test data is a repeat
}
