% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_predict.R
\name{append_to_test}
\alias{append_to_test}
\title{Append fitted item parameters to the test data #
Currently only supports 2PL GPCM - extend for other model types later}
\usage{
append_to_test(test_data, item_params)
}
\arguments{
\item{test_data}{the test data}

\item{item_params}{the fitted item parameters}
}
\value{
the test data with appended item parameters
}
\description{
given a set of test data, and a list of item parameters
append the item parameters for the relevant item in each row to that row
}
