% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bkt.R
\name{bkt_predictions}
\alias{bkt_predictions}
\title{Fit a BKT model and return predicts of the test set}
\usage{
bkt_predictions(data, model)
}
\arguments{
\item{data}{A data frame output from \code{clean_single_response()},
containing an \code{is_test} column}

\item{model}{A BKT model object created using \code{BKT::bkt()}}
}
\value{
A data frame of test set rows with predicted probabilities
}
\description{
Fits a Bayesian Knowledge Tracing (BKT) model to the training data,
returning predictions on the test set.
}
