% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_tune.R
\name{change_pattern}
\alias{change_pattern}
\title{Change Pattern function}
\usage{
change_pattern(
  pattern,
  disc_mean,
  disc_var,
  diff1_mean,
  diff1_var,
  diff2_mean,
  diff2_var
)
}
\arguments{
\item{pattern}{The original pattern to be modified}

\item{disc_mean}{hyper-parameter for the mean of the discrimination parameter's prior}

\item{disc_var}{hyper-parameter for the variance of the discrimination parameter's prior}

\item{diff1_mean}{hyper-parameter for the mean of the first difficulty parameter's parameter's prior}

\item{diff1_var}{hyper-parameter for the variance of the first difficulty parameter's parameter's prior}

\item{diff2_mean}{hyper-parameter for the mean of the second difficulty parameter's parameter's prior}

\item{diff2_var}{hyper-parameter for the variance of the second difficulty parameter's parameter's prior}
}
\value{
The modified response pattern with prior distributions specified from inputted hyper-parameters
}
\description{
Changes the pattern to be used when fitting the IRT mode.
Allows for prior distributions to be set to be set with specific hyper-parameters
}
