% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_all_probs}
\alias{check_all_probs}
\title{Check all test problems appear in training data}
\usage{
check_all_probs(split_data)
}
\arguments{
\item{split_data}{A data frame output from \code{clean_single_response()}}
}
\value{
Logical, TRUE if all test problems appear in training
}
\description{
For IRT models, item parameters can only be estimated for problems seen
during training. This function checks all problems in the test set also
appear in the training set.
}
