% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_all_probs_val}
\alias{check_all_probs_val}
\title{Check all validation problems appear in the training data}
\usage{
check_all_probs_val(split_data)
}
\arguments{
\item{split_data}{A data frame output from \code{clean_single_response()}}
}
\value{
Logical, TRUE if all validation problems appear in training
}
\description{
For IRT models, item parameters can only be estimated for problems seen
during training. This function checks all problems in the validation set also
appear in the training set.
}
