% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{clean_single_response}
\alias{clean_single_response}
\title{Remove students with only one response for a skill}
\usage{
clean_single_response(split_data)
}
\arguments{
\item{split_data}{A data frame output from \code{test_split()}}
}
\value{
A cleaned data frame
}
\description{
Students with only one response will have no training data after the
leave-last-out split. This function removes those students entirely.
}
