% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_fit.R
\name{collapse_train}
\alias{collapse_train}
\title{Addresses repeated interactions between students and items}
\usage{
collapse_train(train_long)
}
\arguments{
\item{train_long}{The training data in long format}
}
\value{
The training data collapsed, so repeat attempts are not included
}
\description{
The training data will contain repeated interactions between the same student and same item.
The collapse_train function will address these observations, combining all interactions between the same items and students
When a student always answers correctly they receive entry 2. When they got it correct at least once, they receive entry 1.
If never correct entry is 0,
}
