% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{count_responses}
\alias{count_responses}
\title{Check minimum number of responses for a skill}
\usage{
count_responses(split_data, minimum)
}
\arguments{
\item{split_data}{A data frame output from \code{clean_single_response()}}

\item{minimum}{An integer giving the minimum number of responses required}
}
\value{
Logical, TRUE if the number of responses meets the minimum
}
\description{
Skills with very few responses may result in empty training sets.
This function checks the total number of responses exceeds a minimum.
}
