% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_predict.R
\name{extract_item_info}
\alias{extract_item_info}
\title{Return all item parameter estimates for items in the training data}
\usage{
extract_item_info(mod)
}
\arguments{
\item{mod}{the fitted IRT model (to the training data)}
}
\value{
the fitted item parameters
}
\description{
given a fitted IRT model, extract all item parameters
}
