% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{extract_skill}
\alias{extract_skill}
\title{Extract data for a single skill}
\usage{
extract_skill(data, skill)
}
\arguments{
\item{data}{A data frame of the full ASSISTments dataset}

\item{skill}{A string giving the skill name to extract}
}
\value{
A data frame filtered to the specified skill
}
\description{
Extract data for a single skill
}
