% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_fit.R
\name{fit_irt}
\alias{fit_irt}
\title{Fits an IRT model where no pattern is specified, but type of IRT model is}
\usage{
fit_irt(data, itemtype = "gpcm")
}
\arguments{
\item{data}{A data frame output from \code{clean_single_response()},
containing an \code{is_test} column}

\item{itemtype}{The specified type of IRT model, default GPCM}
}
\value{
The fitted IRT model
}
\description{
Fits an IRT model using mirt package given an input matrix and specified model type
}
