% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_fit.R
\name{fit_irt_pattern}
\alias{fit_irt_pattern}
\title{Fits an IRT model with a specified pattern}
\usage{
fit_irt_pattern(data, pattern, itemtype = "gpcm")
}
\arguments{
\item{data}{A data frame output from \code{clean_single_response()},
containing an \code{is_test} column}

\item{pattern}{A specified pattern for an IRT model}

\item{itemtype}{The specified type of IRT model, default GPCM}
}
\value{
The fitted IRT model
}
\description{
Fits an IRT model using mirt package given an input matrix and specified pattern
}
