% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_fit.R
\name{fit_irt_pattern_matrix}
\alias{fit_irt_pattern_matrix}
\title{Fits an IRT model with a specified pattern in matrix form}
\usage{
fit_irt_pattern_matrix(irt_matrix, pattern, itemtype = "gpcm")
}
\arguments{
\item{irt_matrix}{A matrix output to fit IRT model to}

\item{pattern}{A specified pattern for an IRT model}

\item{itemtype}{The specified type of IRT model, default GPCM}
}
\value{
The fitted IRT model
}
\description{
Fits an IRT model using mirt package given an input matrix (in matrix form) and specified pattern
}
