% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{has_problem_variation_val}
\alias{has_problem_variation_val}
\title{Check all validation problems have variation in response}
\usage{
has_problem_variation_val(split_data)
}
\arguments{
\item{split_data}{A data frame output from \code{clean_single_response()}}
}
\value{
Logical, TRUE if all validation problems have variation in responses
}
\description{
For IRT models, each problem must have at least one correct and one
incorrect response in the training set for item parameters to be estimated.
}
