% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_predict.R
\name{irt_predictions}
\alias{irt_predictions}
\title{Prediction function for making predictions on the test data given an IRT model}
\usage{
irt_predictions(data, mod)
}
\arguments{
\item{data}{all data with a specified split}

\item{mod}{the fitted IRT model (to the training data)}
}
\value{
the test data with an appended prediction
}
\description{
given a fitted IRT model and the data it was fitted to, makes predictions on the test data
}
