% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_predict.R
\name{make_irt_predictions}
\alias{make_irt_predictions}
\title{Makes IRT predictions on the test data}
\usage{
make_irt_predictions(test_long)
}
\arguments{
\item{test_long}{the test data in long format}
}
\value{
the test data with an extra predicted column
}
\description{
Makes IRT predictions on the test data given we have the information about item parameters and ability
in this test data
}
