% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{test_split}
\alias{test_split}
\title{Split data into training and test sets using leave-last-out}
\usage{
test_split(data)
}
\arguments{
\item{data}{A data frame for a single skill}
}
\value{
A data frame with an additional logical column \code{is_test}
}
\description{
For each student, their final response (based on maximum order_id) is
placed in the test set. All other responses form the training set.
}
