% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_fit.R
\name{train_wide_matrix}
\alias{train_wide_matrix}
\title{Converts the collapsed training data into wide format}
\usage{
train_wide_matrix(train_wide)
}
\arguments{
\item{train_wide}{The collapsed training data in wide format}
}
\value{
The training data in matrix form
}
\description{
To fit IRT models in the mirt package, the data we fit our model to must be in wide, matrix format.
The train_wide_matrix function converts the collapsed wider training data to a matrix format
}
