% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_fit.R
\name{train_wider}
\alias{train_wider}
\title{Converts the collapsed training data into wide format}
\usage{
train_wider(collapsed_train)
}
\arguments{
\item{collapsed_train}{The collapsed training data}
}
\value{
The collapsed training data in wide format
}
\description{
To fit IRT models in the mirt package, the data we fit our model to must be in wide format.
The train_wider function converts the collapsed training data to the wide format
}
