% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_tune.R
\name{loss_tune_irt}
\alias{loss_tune_irt}
\title{Calculate losses on the tuning grid}
\usage{
loss_tune_irt(validate_split, param_grid, itemtype = "gpcm", bound = 0)
}
\arguments{
\item{validate_split}{the validation split of the data}

\item{param_grid}{set of tuning parameters in grid}

\item{itemtype}{the type of IRT model which is fitted, default as gpcm.}

\item{bound}{the bound on the probabilities when using the logarithmic scoring rule
default set at 0}
}
\value{
the parameter grid with the associated loss for each combination of hyper-parameters
when fitted to the validation data
}
\description{
For a given tuning grid, and a validation split of the data
calculates the loss on each combination of hyper-parameters in the tuning grid
}
